:white_check_mark: add comprehensive tests for ObjC bridging, error handling, filters, sorting, and async callbacks
